refactor(lot-journey): extract data builders into utils/lotJourneyData
